Separated `Node` into separate files, moved code to `src` folder
